♻️ refactor: domain/entity -> domain/model 로 네이밍 변경
DB 엔티티와 구분을 위해 도메인 엔티티를 도메인 모델로 네이밍을 변경합니다.
