+++ domain/src/main/kotlin/io/github/jehwanyoo/redis_1st/repository/MovieRepository.kt
@@ -1,6 +1,6 @@
 package io.github.jehwanyoo.redis_1st.repository
 
-import io.github.jehwanyoo.redis_1st.entity.Movie
+import io.github.jehwanyoo.redis_1st.model.Movie
 
 interface MovieRepository {
     fun findAll(): List<Movie>
